requirements.txt
python==3.10
setuptools
pybind11

numpy
scipy
pytorch
